Move the UI-test fixture seeding into its own extension file
The ~1,100 lines of LOCUS_UI_TESTING_* fixture seeding (seedUITestState and
its run/overview fixture helpers) are write-only test plumbing called from one
init site; they now live in AppModel+UITestFixtures.swift. The private(set)
members the seeding writes are widened to internal with an
'internal(for: AppModel+UITestFixtures)' breadcrumb; they re-tighten as each
member moves behind a feature model's restore/ingest seam in later commits.

LocusUITests verified: the transcript reading-column failure reproduces with
identical values on unmodified origin/main on this machine, and the files-tab
filter test passes in isolation, so both are environmental, not regressions.
